Edit the file according to flake8 and eslint
