+++ backend/my_queue/forms.py
@@ -1,9 +1,7 @@
 """Provide forms using in app."""
 
-from django import forms
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
-from .models import Business
 
 
 class SignUpForm(UserCreationForm):
