+++ backend/my_queue/models.py
@@ -1,7 +1,5 @@
 """Provide models using in business app."""
 
-from django import forms
-from django.contrib.auth.forms import UserCreationForm
 from django.db import models
 from django.contrib.auth.models import User
 from django.utils import timezone
