+++ frontend/app/api/logout/route.tsx
@@ -3,7 +3,7 @@
 import { deleteTokens } from "@/lib/auth";
 import { NextResponse } from "next/server";
 
-export async function POST(request: Request): Promise<NextResponse> {
+export async function POST(): Promise<NextResponse> {
     await deleteTokens(); 
     return NextResponse.json({}, { status: 200 }); 
 }
