+++ frontend/app/api/queue/route.tsx
@@ -9,7 +9,7 @@ interface ErrorResponse {
     error: string;
 }
 
-export async function GET(request: Request) {
+export async function GET() {
     try {
         const { data, status }: { data: any; status: number } = await ApiProxy.get(DJANGO_API_QUEUE_URL, true);
         return NextResponse.json<any>(data, { status });
