+++ frontend/app/business/components/BusinessNavbar.tsx
@@ -1,12 +1,10 @@
 'use client'
 
 import React from 'react'
-import { usePathname } from 'next/navigation'
 import Link from 'next/link'
 import Image from 'next/image'
 
 export default function BusinessNavbar() {
-    const pathname = usePathname()
 
     return (
         <div className='navbar fixed z-50 bg-darkPurple max-w-full'>
