+++ frontend/package.json
@@ -9,8 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "jshint": "^2.13.6",
     "next": "15.0.2",
-    "peer": "^1.0.2",
     "primereact": "^10.8.4",
     "qrcode": "^1.5.4",
     "react": "19.0.0-rc-02c0e824-20241028",
@@ -22,9 +22,12 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@typescript-eslint/eslint-plugin": "^8.13.0",
+    "@typescript-eslint/parser": "^8.13.0",
     "daisyui": "^4.12.14",
-    "eslint": "^8",
+    "eslint": "^8.57.1",
     "eslint-config-next": "15.0.2",
+    "peer": "^1.0.2",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
